fix(Package,Config): Added example for PLAYER_NAME prompt
Also prettied
