+++ package.json
@@ -97,6 +97,6 @@
     "discord",
     "discord",
     "rich",
-    "experience"
+    "presence"
   ]
 }
